chore: move `calcLoopedSlides` to shared
