+++ src/solid/loop.js
@@ -1,24 +1,4 @@
-import Swiper from 'swiper';
-
-function calcLoopedSlides(slides, swiperParams) {
-  let slidesPerViewParams = swiperParams.slidesPerView;
-  if (swiperParams.breakpoints) {
-    const breakpoint = Swiper.prototype.getBreakpoint(swiperParams.breakpoints);
-    const breakpointOnlyParams =
-      breakpoint in swiperParams.breakpoints ? swiperParams.breakpoints[breakpoint] : undefined;
-    if (breakpointOnlyParams && breakpointOnlyParams.slidesPerView) {
-      slidesPerViewParams = breakpointOnlyParams.slidesPerView;
-    }
-  }
-  let loopedSlides = Math.ceil(parseFloat(swiperParams.loopedSlides || slidesPerViewParams, 10));
-
-  loopedSlides += swiperParams.loopAdditionalSlides;
-
-  if (loopedSlides > slides.length && swiperParams.loopedSlidesLimit) {
-    loopedSlides = slides.length;
-  }
-  return loopedSlides;
-}
+import { calcLoopedSlides } from '../shared/calc-looped-slides.js';
 
 function renderLoop(swiper, slides, swiperParams) {
   const modifiedSlides = slides.map((child, index) => {
